ensure that uploading from within directories always uploads to that directory
Unfortunately when you hit upload we only get the current path, and have trouble telling whether you're viewing a directory or a file. (ember architectural decisions which will need to be cleaned up later) So we need to ensure that directories end in a / so that we can tell them apart.