Disable javascript in iframe preview
My (mis)understanding was that an iframe sandbox that lacked "allow-same-origin" would have no document.location and would not be able to resolve relative URLs. The first is true, the second is not. So it is possible to XHR POST from the iframe sandbox, and while you can't read the response due to cross-origin, you can create files and delete files. We're disabling javascript inside html previews until we can come up with a better solution.