Add properties needed by OwnCloud and NextCloud
OwnCloud and NextCloud requires that the properties
`{http://owncloud.org/ns}id` and `{http://owncloud.org/ns}permissions`
are present on all files and folders so I have simply added them with
the most simple implementation.

Id is simply inode in string form and permissions is a static text for
files and folders.

Closes: #93